productivity increasing
preprocessing and taking derivative and boundary calc are changed

+++ examples/example_Painleve_I.py
@@ -65,14 +65,14 @@ def p_I_exp(grid_res,nruns,CACHE):
     """
     
     # point t=0
-    bnd1 = torch.from_numpy(np.array([[0]], dtype=np.float64))
+    bnd1 = torch.from_numpy(np.array([[0]], dtype=np.float64)).float()
     
     
     #  So u(0)=-1/2
     bndval1 = torch.from_numpy(np.array([[0]], dtype=np.float64))
     
     # point t=1
-    bnd2 = torch.from_numpy(np.array([[float(t[0])]], dtype=np.float64))
+    bnd2 = torch.from_numpy(np.array([[float(t[0])]], dtype=np.float64)).float()
     
     # d/dt
     bop2 ={
